feat: integrating with aur
