+++ internal/pipe/aur/aur.go
@@ -5,7 +5,6 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"sort"
@@ -16,17 +15,15 @@ import (
 	"github.com/goreleaser/goreleaser/internal/artifact"
 	"github.com/goreleaser/goreleaser/internal/client"
 	"github.com/goreleaser/goreleaser/internal/commitauthor"
-	"github.com/goreleaser/goreleaser/internal/git"
 	"github.com/goreleaser/goreleaser/internal/pipe"
 	"github.com/goreleaser/goreleaser/internal/tmpl"
 	"github.com/goreleaser/goreleaser/pkg/config"
 	"github.com/goreleaser/goreleaser/pkg/context"
-	"golang.org/x/crypto/ssh"
 )
 
 const (
 	aurExtra          = "AURConfig"
-	defaultSSHCommand = "ssh -i {{ .KeyPath }} -o StrictHostKeyChecking=accept-new -F /dev/null"
+	defaultSSHCommand = client.DefaulGittSSHCommand
 	defaultCommitMsg  = "Update to {{ .Tag }}"
 )
 
@@ -62,7 +59,7 @@ func (Pipe) Default(ctx *context.Context) error {
 			pkg.Rel = "1"
 		}
 		if pkg.GitSSHCommand == "" {
-			pkg.GitSSHCommand = defaultSSHCommand
+			pkg.GitSSHCommand = client.DefaulGittSSHCommand
 		}
 		if pkg.Goamd64 == "" {
 			pkg.Goamd64 = "v1"
@@ -368,149 +365,28 @@ func doPublish(ctx *context.Context, pkgs []*artifact.Artifact) error {
 		return pipe.Skip("prerelease detected with 'auto' upload, skipping aur publish")
 	}
 
-	key, err := tmpl.New(ctx).Apply(cfg.PrivateKey)
+	cli, err := client.NewGitUploadClient(ctx)
 	if err != nil {
 		return err
 	}
 
-	key, err = keyPath(key)
-	if err != nil {
-		return err
-	}
+	repo := client.RepoFromRef(config.RepoRef{
+		PrivateKey:    cfg.PrivateKey,
+		GitURL:        cfg.GitURL,
+		GitSSHCommand: cfg.GitSSHCommand,
+	})
 
-	url, err := tmpl.New(ctx).Apply(cfg.GitURL)
-	if err != nil {
-		return err
-	}
-
-	if url == "" {
-		return pipe.Skip("aur.git_url is empty")
-	}
-
-	sshcmd, err := tmpl.New(ctx).WithExtraFields(tmpl.Fields{
-		"KeyPath": key,
-	}).Apply(cfg.GitSSHCommand)
-	if err != nil {
-		return err
-	}
-
-	msg, err := tmpl.New(ctx).Apply(cfg.CommitMessageTemplate)
-	if err != nil {
-		return err
-	}
-
-	author, err := commitauthor.Get(ctx, cfg.CommitAuthor)
-	if err != nil {
-		return err
-	}
-
-	parent := filepath.Join(ctx.Config.Dist, "aur", "repos")
-	cwd := filepath.Join(parent, cfg.Name)
-
-	if err := os.MkdirAll(parent, 0o755); err != nil {
-		return err
-	}
-
-	env := []string{fmt.Sprintf("GIT_SSH_COMMAND=%s", sshcmd)}
-
-	if err := runGitCmds(ctx, parent, env, [][]string{
-		{"clone", url, cfg.Name},
-	}); err != nil {
-		return fmt.Errorf("failed to setup local AUR repo: %w", err)
-	}
-
-	if err := runGitCmds(ctx, cwd, env, [][]string{
-		// setup auth et al
-		{"config", "--local", "user.name", author.Name},
-		{"config", "--local", "user.email", author.Email},
-		{"config", "--local", "commit.gpgSign", "false"},
-		{"config", "--local", "init.defaultBranch", "master"},
-	}); err != nil {
-		return fmt.Errorf("failed to setup local AUR repo: %w", err)
-	}
-
+	log.WithField("repo", cfg.GitURL).WithField("name", cfg.Name).Info("pushing")
 	for _, pkg := range pkgs {
-		bts, err := os.ReadFile(pkg.Path)
+		log.WithField("package", pkg.Name).Info("pushing")
+		content, err := os.ReadFile(pkg.Path)
 		if err != nil {
-			return fmt.Errorf("failed to read %s: %w", pkg.Name, err)
+			return err
 		}
-
-		if err := os.WriteFile(filepath.Join(cwd, pkg.Name), bts, 0o644); err != nil {
-			return fmt.Errorf("failed to write %s: %w", pkg.Name, err)
-		}
-	}
-
-	log.WithField("repo", url).WithField("name", cfg.Name).Info("pushing")
-	if err := runGitCmds(ctx, cwd, env, [][]string{
-		{"add", "-A", "."},
-		{"commit", "-m", msg},
-		{"push", "origin", "HEAD"},
-	}); err != nil {
-		return fmt.Errorf("failed to push %q (%q): %w", cfg.Name, url, err)
-	}
-
-	return nil
-}
-
-func keyPath(key string) (string, error) {
-	if key == "" {
-		return "", pipe.Skip("aur.private_key is empty")
-	}
-
-	path := key
-
-	_, err := ssh.ParsePrivateKey([]byte(key))
-	if isPasswordError(err) {
-		return "", fmt.Errorf("key is password-protected")
-	}
-
-	if err == nil {
-		// if it can be parsed as a valid private key, we write it to a
-		// temp file and use that path on GIT_SSH_COMMAND.
-		f, err := os.CreateTemp("", "id_*")
-		if err != nil {
-			return "", fmt.Errorf("failed to store private key: %w", err)
-		}
-		defer f.Close()
-
-		// the key needs to EOF at an empty line, seems like github actions
-		// is somehow removing them.
-		if !strings.HasSuffix(key, "\n") {
-			key += "\n"
-		}
-
-		if _, err := io.WriteString(f, key); err != nil {
-			return "", fmt.Errorf("failed to store private key: %w", err)
-		}
-		if err := f.Close(); err != nil {
-			return "", fmt.Errorf("failed to store private key: %w", err)
-		}
-		path = f.Name()
-	}
-
-	if _, err := os.Stat(path); err != nil {
-		return "", fmt.Errorf("could not stat aur.private_key: %w", err)
-	}
-
-	// in any case, ensure the key has the correct permissions.
-	if err := os.Chmod(path, 0o600); err != nil {
-		return "", fmt.Errorf("failed to ensure aur.private_key permissions: %w", err)
-	}
-
-	return path, nil
-}
-
-func isPasswordError(err error) bool {
-	var kerr *ssh.PassphraseMissingError
-	return errors.As(err, &kerr)
-}
-
-func runGitCmds(ctx *context.Context, cwd string, env []string, cmds [][]string) error {
-	for _, cmd := range cmds {
-		args := append([]string{"-C", cwd}, cmd...)
-		if _, err := git.Clean(git.RunWithEnv(ctx, env, args...)); err != nil {
-			return fmt.Errorf("%q failed: %w", strings.Join(cmd, " "), err)
+		if err := cli.CreateFile(ctx, cfg.CommitAuthor, repo, content, pkg.Name, cfg.CommitMessageTemplate); err != nil {
+			return err
 		}
 	}
+
 	return nil
 }
